--- BlockType.java
@@ -1,5 +1,0 @@
-package Snake;
-
-public enum BlockType {
-	BLANK, BODY, HEAD, APPLE, WALL;
-}
