+++ Snake/BlockType.java
@@ -1,0 +1,5 @@
+package Snake;
+
+public enum BlockType {
+	BLANK, BODY, HEAD, APPLE, WALL;
+}
