+++ Snake/snakeskin/blank.png
