+++ Snake/snakeskin/wall1.png
