+++ Snake/snakeskin/wall2.png
